+++ src/types/chain.ts
@@ -1,0 +1,17 @@
+export enum ChainType {
+  'evm',
+  'solana'
+}
+
+export enum ChainId {
+  'ethereum',
+  'polygon',
+  'solana'
+}
+
+export type Chain = {
+  id: ChainId;
+  name: string;
+  rpcUrl: string;
+  type: ChainType
+}
