+++ src/types/nft.ts
@@ -1,6 +1,7 @@
 import { ValidContractNFTCallFunction } from '../clients/ethereum';
 
 import { ArtistProfile } from './artist';
+import { ChainId } from './chain';
 import { Contract } from './contract';
 import { ExtractorTypes } from './fieldExtractor';
 import { NFTFactoryTypes } from './nftFactory';
@@ -69,6 +70,7 @@ export type NftFactory = Contract & {
   standard: NFTStandard
   autoApprove: boolean; // automaticlly approve nfts generated by this facotry
   approved: boolean; // should index nfts generated by this factory
+  chainId: ChainId;
 }
 
 export type NFTContractType = {
